+++ OP/Cpreprocessor/src/assets/utils.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdlib.h>
 #define MAX_ALLOC_ATTEMPTS 10000
 
-void* myAllocMemory(unsigned int size);
+void* myAllocMemory(size_t size);
